requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
Pillow==12.0.0